+++ db-scheduler-ui-frontend/src/components/RefreshButton.tsx
@@ -1,6 +1,6 @@
 import { Box, Button } from '@chakra-ui/react';
 import { RepeatIcon } from '@chakra-ui/icons';
-import React, { useState } from 'react';
+import React from 'react';
 import colors from 'src/styles/colors';
 import {
   InfiniteData,
@@ -9,8 +9,8 @@ import {
 } from '@tanstack/react-query';
 import { POLL_TASKS_QUERY_KEY, pollTasks } from 'src/services/pollTasks';
 import { TaskDetailsRequestParams } from 'src/models/TaskRequestParams';
-import { NumberCircle } from './NumberCircle';
 import { TasksResponse } from 'src/models/TasksResponse';
+import { RefreshCircle } from './RefreshCircle';
 
 interface RefreshButtonProps {
   refetch?: () => Promise<
@@ -91,59 +91,6 @@ export const RefreshButton: React.FC<RefreshButtonProps> = ({
           textColor={colors.primary['900']}
           visible={data?.newTasks !== 0}
           hoverText=" added since refresh"
-        />
-      </Box>
-    </Box>
-  );
-};
-
-type RefreshCircleProps = {
-  number: number;
-  color: string;
-  textColor?: string;
-  visible?: boolean;
-  hoverText: string;
-};
-
-const RefreshCircle: React.FC<RefreshCircleProps> = ({
-  number,
-  color,
-  textColor,
-  visible,
-  hoverText,
-}) => {
-  const [hovered, setHovered] = useState(false);
-
-  const text = hovered ? hoverText : '';
-
-  const powerOfTen = (number + hoverText).length - 1;
-  const isExpanded = 1 <= powerOfTen;
-  const baseSize: number = 22;
-  const width = isExpanded ? baseSize + 7 * powerOfTen : baseSize;
-
-  const marginLeft = isExpanded ? -width : 0;
-
-  return (
-    <Box
-      alignItems={'end'}
-      display={'flex'}
-      justifyContent={'flex-end'}
-      overflow={'visible'}
-      ml={marginLeft}
-      position="relative"
-      visibility={visible ? 'visible' : 'hidden'}
-    >
-      <Box
-        onMouseEnter={() => setHovered(true)}
-        onMouseLeave={() => setHovered(false)}
-      >
-        <NumberCircle
-          number={number + text}
-          bgColor={color}
-          textColor={textColor}
-          position="relative"
-          top={'auto'}
-          style={{ bottom: '0', left: '0' }}
         />
       </Box>
     </Box>
