+++ db-scheduler-ui-frontend/src/services/getAllLogs.ts
@@ -1,4 +1,4 @@
-import { FilterBy } from './getTasks';
+import { FilterBy } from 'src/models/QueryParams';
 
 const API_BASE_URL: string =
   (import.meta.env.VITE_API_BASE_URL as string) ??
